Renaming KSCPU variants
Renaming KSCPU variants
