+++ Sources/KSCrash/Recording/Tools/Raygun_KSCPU_arm.c
@@ -28,7 +28,7 @@
 
 
 #include "Raygun_KSCPU.h"
-#include "KSCPU_Apple.h"
+#include "Raygun_KSCPU_Apple.h"
 #include "KSMachineContext.h"
 #include "KSMachineContext_Apple.h"
 #include <stdlib.h>
